[OrderService] Moving directory files/packages
