+++ OrderService/Infrastructure/Data/Mappings/CreditCardMapping.cs
@@ -2,7 +2,7 @@
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
 
-namespace Infrastructure.Data.ModelsMappings
+namespace Infrastructure.Data.Mappings
 {
     public class CreditCardMapping : IEntityTypeConfiguration<CreditCard>
     {
