+++ OrderService/Infrastructure/Data/Mappings/PaymentMapping.cs
@@ -2,7 +2,7 @@
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
 
-namespace Infrastructure.Data.ModelsMappings
+namespace Infrastructure.Data.Mappings
 {
     public class PaymentMapping : IEntityTypeConfiguration<Payment>
     {
